expense_agent_app/requirements.txt
openai>=1.40.0
gradio>=4.44.0
python-dotenv>=1.0.0
requests>=2.32.0
